Fix exception raising in pool methods
